[18.0][NEW] Module 'oca_membership_groups', replacement of 'oca_custom': 1. Working Groups boolean, "Groups" entry menu in "Members" 2. Keep up-to-date Mailing Groups as per member's Membership Categories (Member & Delegate) 3. When a portal member unsubscribe from a Mail Group, keep its choice instead of unlinking    the membership, to prevent spamming him/her 4. Grace period when a member should be removed from a Mail Group (configurable, per Mail    Group)
[IMP] Bridge with Email Marketing: allow to send marketing mail using Mail Groups Members

[IMP] Simplify behavior of 'membership_category_ids' for invoicing + Detbors management (new filters & editable tree view)

+++ oca_membership_groups/models/mail_group_member.py
@@ -2,9 +2,8 @@
 # @author Arnaud LAYEC <arnaud.layec@akretion.com>
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
-from odoo import fields, models, api, exceptions, _
+from odoo import fields, models, api, tools
 from datetime import timedelta
-
 
 class MailGroupMember(models.Model):
     _inherit = ["mail.group.member"]
@@ -12,32 +11,15 @@ class MailGroupMember(models.Model):
     active = fields.Boolean(default=True)
     grace_date_start = fields.Date(
         string="Grace Date Start",
-        readonly=True,
         help="Date on which the member is not legitimate anymore to belong "
-             "in this Mailing"
+             "in this Mailing",
     )
     grace_date_deadline = fields.Date(
-        string="Grace Date Deadline",
+        string="Grace Deadline",
         compute="_compute_grace_date_deadline",
         help="Date on which the member will be retired automatically from this "
              "Mailing Group, unless he/she renew its membership.",
     )
-
-    #===== Constrain =====#
-    @api.constrains("mail_group_id")
-    def _constrain_mail_group_ids(self):
-        if self._context.get("allow_membership_provision_groups"):
-            return
-        members = self._get_auto_member()
-        if members:
-            raise exceptions.UserError(_(
-                "Modifying members of the following Mail Groups cannot be done "
-                "manually, because those groups follow member's roles: %s.",
-                ", " . join(members.mapped("name"))
-            ))
-
-    def _get_auto_member(self):
-        return self.mail_group_id.filtered("membership_category_ids")
 
     #===== Compute =====#
     @api.depends("grace_date_start", "mail_group_id.grace_days")
@@ -45,16 +27,21 @@ class MailGroupMember(models.Model):
         for member in self:
             member.grace_date_deadline = (
                 bool(member.grace_date_start) and
-                member.grace_date_start + timedelta(days=member.mail_group_id.grace_day)
+                member.grace_date_start + timedelta(days=member.mail_group_id.grace_days)
             )
 
     #===== CRUD =====#
-    @api.model_create_multi
-    def create(self, vals_list):
-        records = super().create(vals_list)
-        records._constrain_mail_group_ids()
-        return records
+    def write(self, vals):
+        """When called from `_join_group`, if a user re-subscribe a previously
+        unsubscribed group, re-active its membership instead of creating a duplicate"""
+        if self._context.get("from_portal"):
+            vals["active"] = True
+        return super().write(vals)
 
     def unlink(self):
-        self._constrain_mail_group_ids()
-        return super().unlink()
+        """When a user leaves a group, remember its unsubscription
+        (=archive instead of unlink) to avoid provisioning it automatically"""
+        if self._context.get("from_portal"):
+            self.with_context(from_portal=False).action_archive()
+        else:
+            return super().unlink()
